Add missing byte order handling for geometry data
Add missing byte order handling for geometry data. Check whether
the current platform is LE or BE and then swizzle the data around
if the byte order differs from the data's byte order.

+++ graphics/unit_test/unit_test_drawable.cpp
@@ -241,7 +241,7 @@ void unit_test_wireframe()
         TEST_REQUIRE(wireframe.GetNumDrawCmds() == 1);
         TEST_REQUIRE(wireframe.GetDrawCmd(0).type == gfx::Geometry::DrawType::Lines);
         TEST_REQUIRE(wireframe.GetDrawCmd(0).offset == 0);
-        TEST_REQUIRE(wireframe.GetDrawCmd(0).count == std::numeric_limits<size_t>::max());
+        TEST_REQUIRE(wireframe.GetDrawCmd(0).count == std::numeric_limits<uint32_t>::max());
 
         const gfx::VertexStream stream(wireframe.GetLayout(),
                                  wireframe.GetVertexDataPtr(),
@@ -283,7 +283,7 @@ void unit_test_wireframe()
         TEST_REQUIRE(wireframe.GetNumDrawCmds() == 1);
         TEST_REQUIRE(wireframe.GetDrawCmd(0).type == gfx::Geometry::DrawType::Lines);
         TEST_REQUIRE(wireframe.GetDrawCmd(0).offset == 0);
-        TEST_REQUIRE(wireframe.GetDrawCmd(0).count == std::numeric_limits<size_t>::max());
+        TEST_REQUIRE(wireframe.GetDrawCmd(0).count == std::numeric_limits<uint32_t>::max());
 
         const gfx::VertexStream stream(wireframe.GetLayout(),
                                        wireframe.GetVertexDataPtr(),
